seekBusiness: Inherit from apac theme rather than seekAnz
