+++ .changeset/three-pens-collect.md
@@ -1,0 +1,14 @@
+---
+'braid-design-system': minor
+---
+
+**`seekBusiness` theme:** Inherit from new `apac` theme rather than the deprecated `seekAnz` theme
+
+Just like the migration from `seekAnz` to `apac`, the visual changes are as follows:
+
+- The body background has changed from `#eeeeee` to `#f5f6f8`.
+- All grey colours now have a hint of blue.
+- Buttons and form fields have decreased in height from `48px` to `44px`.
+- Border radiuses have increased from `2px` to `4px`.
+
+It's possible that your application has custom design elements that depend on these older values in order to look correct. It's also possible that your application is part of a broader user journey that makes use of these older design standards. As a result, a **design review is highly recommended.**
